test(toast): migrate standalone
