+++ core/src/components/toast/test/standalone/toast.e2e.ts-snapshots/toast-standalone-md-ltr-Mobile-Firefox-linux.png
